requirements.txt.txt
streamlit
langchain-community
langchain-text-splitters
langchain-openai
langchain-core
pypdf
chromadb
sentence-transformers
python-pptx
python-docx